rename Search to SearchInput
to better signify this is a variant/extension of the Input component

we will have other similar components like PasswordInput etc.
