+++ apps/docs/demos/search-input/disabled-usage/App.tsx
@@ -1,0 +1,5 @@
+import { SearchInput } from "@optiaxiom/react";
+
+export function App() {
+  return <SearchInput defaultValue="Some disabled value" disabled w="240" />;
+}
