+++ apps/docs/demos/search-input/usage/App.tsx
@@ -1,0 +1,5 @@
+import { SearchInput } from "@optiaxiom/react";
+
+export function App() {
+  return <SearchInput placeholder="Enter text..." w="240" />;
+}
