--- apps/docs/demos/search/disabled-usage/App.tsx
@@ -1,5 +1,0 @@
-import { Search } from "@optiaxiom/react";
-
-export function App() {
-  return <Search defaultValue="Some disabled value" disabled w="240" />;
-}
