--- apps/docs/demos/search/usage/App.tsx
@@ -1,5 +1,0 @@
-import { Search } from "@optiaxiom/react";
-
-export function App() {
-  return <Search placeholder="Enter text..." w="240" />;
-}
