+++ apps/docs/pages/components/search-input.mdx
@@ -1,4 +1,4 @@
-# Search
+# SearchInput
 
 Basic search input field with clear button.
 
@@ -6,25 +6,25 @@ Basic search input field with clear button.
 
 ### Usage
 
-The `Search` component extends the `Input` component with some additional decorators.
+The `SearchInput` component extends the `Input` component with some additional decorators.
 
-<ax-demo name="search/usage" />
+<ax-demo name="search-input/usage" />
 
 ### Disabled state
 
 Enable the `disabled` prop to toggle the disabled state of the input field.
 
-<ax-demo name="search/disabled-usage" />
+<ax-demo name="search-input/disabled-usage" />
 
 ### Controlled
 
 Use the `value` and `defaultValue` props to toggle between controlled and uncontrolled usage. And combine it with `onChange` to listen for changes to the input field.
 
-<ax-demo name="search/controlled-usage" />
+<ax-demo name="search-input/controlled-usage" />
 
 ## Props
 
-<ax-props-table component="Search" extends="Input" />
+<ax-props-table component="SearchInput" extends="Input" />
 
 ## Changelog
