+++ packages/react/src/search-input/index.ts
@@ -1,0 +1,1 @@
+export * from "./SearchInput";
